Adapt renderer for getegrid
