+++ pyramid_oereb/lib/renderer/getegrid/templates/xml/getegrid.xml
@@ -4,9 +4,21 @@
                   xmlns:xsd="http://www.w3.org/2001/XMLSchema"
                   xmlns="http://schemas.geo.admin.ch/V_D/OeREB/1.0/Extract"
                   xsi:schemaLocation="http://schemas.geo.admin.ch/V_D/OeREB/1.0/Extract http://schemas.geo.admin.ch/V_D/OeREB/1.0/Extract.xsd">
-    % for item in data['GetEGRIDResponse']:
+% for item in data['GetEGRIDResponse']:
     <egrid>${item['egrid']}</egrid>
     <number>${item['number']}</number>
     <identDN>${item['identDN']}</identDN>
-    % endfor
+    <type>${item['type']}</type>
+    %if item['limit'] and params.with_geometry:
+    <data:Limit>
+        <gml:MultiSurface gml:id="${get_gml_id()}">
+            %for polygon in item['limit'].geoms:
+            <gml:surfaceMember>
+                <%include file="polygon.xml" args="polygon=polygon"/>
+            </gml:surfaceMember>
+            %endfor
+        </gml:MultiSurface>
+    </data:Limit>
+    %endif
+% endfor
 </GetEGRIDResponse>
